Removed various administrative procedures
Meeting minutes: ain't-nobody-got-time-for-that.jpg
We'll always make a best effort to record sessions.
Otherwise be there or be square.

Meeting promotion: Stating the obvious
Nitty gritty details not useful.

BR 25 requirement: We've never done it, impractical to track.

Staff Meetings only valid if informed 2 days prior:
Staff Meetings are nothing special.
If members don't get to participate, why do they care if a meeting's
going on in the background between staff?
Staff need a bit of notice if an important decision needs to be made.
Otherwise we should be able to meet whenever.